Creation, connection of modules
